Feature : AminPage에서 탭 내부 컨텐츠를컴포넌트로 분리
